+++ bdtsim/cli/main.py
@@ -18,6 +18,7 @@
 from typing import Optional
 
 from .command_manager import CommandManager
+from .environment_info import EnvironmentInfoSubCommand
 from .list_environments import ListEnvironmentsSubCommand
 from .list_protocols import ListProtocolsSubCommand
 from .run import RunSubCommand
@@ -27,5 +28,6 @@ def main() -> Optional[int]:
     command_manager = CommandManager()
     command_manager.register_subcommand('list-protocols', ListProtocolsSubCommand)
     command_manager.register_subcommand('list-environments', ListEnvironmentsSubCommand)
+    command_manager.register_subcommand('environment-info', EnvironmentInfoSubCommand)
     command_manager.register_subcommand('run', RunSubCommand)
     return command_manager.run()